+++ pki/ca.go
@@ -19,11 +19,9 @@ import (
 	"crypto"
 	"crypto/ecdsa"
 	"crypto/elliptic"
-	"crypto/md5"
 	"crypto/rand"
 	"crypto/x509"
 	"crypto/x509/pkix"
-	"encoding/base64"
 	"log"
 	"math/big"
 	rdm "math/rand"
@@ -114,30 +112,26 @@ func (ca *RootCA) CAChain() ([]*x509.Certificate, error) {
 }
 
 // SignCSR signs a ASN.1 DER encoded certificate signing request.
-func (ca *RootCA) SignCSR(der []byte) (*x509.Certificate, error) {
+func (ca *RootCA) SignCSR(der []byte, template *x509.Certificate) (*x509.Certificate, error) {
 	csr, err := x509.ParseCertificateRequest(der)
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to parse CSR")
 	}
-
-	// Certificate CN is base64-encoded (w/o padding) MD5 hash of the public key
-	hash := md5.Sum(csr.RawSubjectPublicKeyInfo)
-	cn := base64.RawURLEncoding.EncodeToString(hash[:])
 
 	// Pick random serial number
 	source := rdm.NewSource(time.Now().UnixNano())
 	serial := big.NewInt(int64(rdm.New(source).Uint64()))
 
 	// Sign certificate request
-	template := &x509.Certificate{
+	tmpl := &x509.Certificate{
 		SerialNumber: serial,
-		Subject:      pkix.Name{CommonName: cn},
+		Subject:      template.Subject,
 		NotBefore:    time.Now(),
 		NotAfter:     ca.Cert.NotAfter, // Valid until CA expires
 	}
 	certDER, err := x509.CreateCertificate(
 		rand.Reader,
-		template,
+		tmpl,
 		ca.Cert,
 		csr.PublicKey,
 		ca.Key,